DataGeneration/src/requirements.txt
alabaster==0.7.12
attrs==22.1.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.2.0
Cython==3.0.0a9
docutils==0.17.1
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==3.0.2
MarkupSafe==2.0.1
networkx==2.6.3
packaging==21.2
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pytest==7.1.2
pytest-mock==3.8.2
python-on-whales==0.31.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.0
urllib3==1.26.7
